# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Burger Nomr - Halaal burger nom platform for Cape Town
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface Restaurant {
  id: string;
  name: string;
  area: string;
  address: string;
  description: string;
  image_url?: string | null;
  avg_score?: number | null;
  total_noms: number;
  price_range?: string | null;
  tags: string[];
  latitude?: number | null;
  longitude?: number | null;
  created_at: string;
}

export interface Nom {
  id: string;
  restaurant_id: string;
  user_id: string;
  user_name: string;
  user_avatar?: string | null;
  score: number;
  comment?: string | null;
  created_at: string;
}

export type RestaurantDetail = Restaurant & {
  noms?: Nom[];
};

export type RestaurantWithRank = Restaurant & {
  rank: number;
};

export interface UserProfile {
  id: string;
  name: string;
  email: string;
  avatar_url?: string | null;
  total_noms: number;
  joined_at: string;
}

export interface SummaryStats {
  total_restaurants: number;
  total_noms: number;
  top_area?: string | null;
  avg_score_platform?: number | null;
}

export interface CreateRestaurantBody {
  name: string;
  area: string;
  address: string;
  description: string;
  image_url?: string | null;
  price_range?: string | null;
  tags: string[];
  latitude?: number | null;
  longitude?: number | null;
}

export interface CreateNomBody {
  restaurant_id: string;
  user_id: string;
  user_name: string;
  user_avatar?: string | null;
  score: number;
  comment?: string | null;
}

export type ListRestaurantsParams = {
  search?: string;
  area?: string;
  limit?: number;
  offset?: number;
};

export type GetRecentRestaurantsParams = {
  limit?: number;
};