# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Burger Nomr - Halaal burger nom platform for Cape Town
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: restaurants
    description: Restaurant operations
  - name: noms
    description: Nom operations
  - name: users
    description: User operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /restaurants:
    get:
      operationId: listRestaurants
      tags: [restaurants]
      summary: List all restaurants
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: area
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of restaurants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Restaurant"
    post:
      operationId: createRestaurant
      tags: [restaurants]
      summary: Submit a new restaurant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRestaurantBody"
      responses:
        "201":
          description: Created restaurant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"

  /restaurants/top:
    get:
      operationId: getTopRestaurants
      tags: [restaurants]
      summary: Get top 10 restaurants by average nom score
      responses:
        "200":
          description: Top 10 restaurants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RestaurantWithRank"

  /restaurants/recent:
    get:
      operationId: getRecentRestaurants
      tags: [restaurants]
      summary: Get recently added restaurants
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 6
      responses:
        "200":
          description: Recently added restaurants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Restaurant"

  /restaurants/{id}:
    get:
      operationId: getRestaurant
      tags: [restaurants]
      summary: Get restaurant by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Restaurant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantDetail"
        "404":
          description: Not found

  /restaurants/{id}/noms:
    get:
      operationId: getRestaurantNoms
      tags: [restaurants]
      summary: Get all noms for a restaurant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Noms for restaurant
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Nom"

  /noms:
    post:
      operationId: createNom
      tags: [noms]
      summary: Submit a nom score
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNomBody"
      responses:
        "201":
          description: Created nom
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Nom"

  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  /users/{id}/noms:
    get:
      operationId: getUserNoms
      tags: [users]
      summary: Get all noms by a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User noms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Nom"

  /stats/summary:
    get:
      operationId: getSummaryStats
      tags: [restaurants]
      summary: Get platform summary stats
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SummaryStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Restaurant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        area:
          type: string
        address:
          type: string
        description:
          type: string
        image_url:
          type: string
          nullable: true
        avg_score:
          type: number
          nullable: true
        total_noms:
          type: integer
        price_range:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        created_at:
          type: string
      required:
        - id
        - name
        - area
        - address
        - description
        - total_noms
        - tags
        - created_at

    RestaurantDetail:
      allOf:
        - $ref: "#/components/schemas/Restaurant"
        - type: object
          properties:
            noms:
              type: array
              items:
                $ref: "#/components/schemas/Nom"

    RestaurantWithRank:
      allOf:
        - $ref: "#/components/schemas/Restaurant"
        - type: object
          properties:
            rank:
              type: integer
          required:
            - rank

    Nom:
      type: object
      properties:
        id:
          type: string
        restaurant_id:
          type: string
        user_id:
          type: string
        user_name:
          type: string
        user_avatar:
          type: string
          nullable: true
        score:
          type: number
        comment:
          type: string
          nullable: true
        created_at:
          type: string
      required:
        - id
        - restaurant_id
        - user_id
        - user_name
        - score
        - created_at

    UserProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        avatar_url:
          type: string
          nullable: true
        total_noms:
          type: integer
        joined_at:
          type: string
      required:
        - id
        - name
        - email
        - total_noms
        - joined_at

    SummaryStats:
      type: object
      properties:
        total_restaurants:
          type: integer
        total_noms:
          type: integer
        top_area:
          type: string
          nullable: true
        avg_score_platform:
          type: number
          nullable: true
      required:
        - total_restaurants
        - total_noms

    CreateRestaurantBody:
      type: object
      properties:
        name:
          type: string
        area:
          type: string
        address:
          type: string
        description:
          type: string
        image_url:
          type: string
          nullable: true
        price_range:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
      required:
        - name
        - area
        - address
        - description
        - tags

    CreateNomBody:
      type: object
      properties:
        restaurant_id:
          type: string
        user_id:
          type: string
        user_name:
          type: string
        user_avatar:
          type: string
          nullable: true
        score:
          type: number
        comment:
          type: string
          nullable: true
      required:
        - restaurant_id
        - user_id
        - user_name
        - score