# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Burger Nomr - Halaal burger nom platform for Cape Town
 * OpenAPI spec version: 0.2.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all restaurants
 */
export const listRestaurantsQueryLimitDefault = 20;
export const listRestaurantsQueryOffsetDefault = 0;

export const ListRestaurantsQueryParams = zod.object({
  search: zod.coerce.string().optional(),
  area: zod.coerce.string().optional(),
  limit: zod.coerce.number().default(listRestaurantsQueryLimitDefault),
  offset: zod.coerce.number().default(listRestaurantsQueryOffsetDefault),
});

export const ListRestaurantsResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  area: zod.string(),
  address: zod.string(),
  description: zod.string(),
  image_url: zod.string().nullish(),
  avg_score: zod.number().nullish(),
  total_noms: zod.number(),
  price_range: zod.string().nullish(),
  tags: zod.array(zod.string()),
  latitude: zod.number().nullish(),
  longitude: zod.number().nullish(),
  created_at: zod.string(),
});
export const ListRestaurantsResponse = zod.array(ListRestaurantsResponseItem);

/**
 * @summary Submit a new restaurant
 */
export const CreateRestaurantBody = zod.object({
  name: zod.string(),
  area: zod.string(),
  address: zod.string(),
  description: zod.string(),
  image_url: zod.string().nullish(),
  price_range: zod.string().nullish(),
  tags: zod.array(zod.string()),
  latitude: zod.number().nullish(),
  longitude: zod.number().nullish(),
});

/**
 * @summary Get top 10 restaurants by average nom score
 */
export const GetTopRestaurantsResponseItem = zod
  .object({
    id: zod.string(),
    name: zod.string(),
    area: zod.string(),
    address: zod.string(),
    description: zod.string(),
    image_url: zod.string().nullish(),
    avg_score: zod.number().nullish(),
    total_noms: zod.number(),
    price_range: zod.string().nullish(),
    tags: zod.array(zod.string()),
    latitude: zod.number().nullish(),
    longitude: zod.number().nullish(),
    created_at: zod.string(),
  })
  .and(
    zod.object({
      rank: zod.number(),
    }),
  );
export const GetTopRestaurantsResponse = zod.array(
  GetTopRestaurantsResponseItem,
);

/**
 * @summary Get recently added restaurants
 */
export const getRecentRestaurantsQueryLimitDefault = 6;

export const GetRecentRestaurantsQueryParams = zod.object({
  limit: zod.coerce.number().default(getRecentRestaurantsQueryLimitDefault),
});

export const GetRecentRestaurantsResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  area: zod.string(),
  address: zod.string(),
  description: zod.string(),
  image_url: zod.string().nullish(),
  avg_score: zod.number().nullish(),
  total_noms: zod.number(),
  price_range: zod.string().nullish(),
  tags: zod.array(zod.string()),
  latitude: zod.number().nullish(),
  longitude: zod.number().nullish(),
  created_at: zod.string(),
});
export const GetRecentRestaurantsResponse = zod.array(
  GetRecentRestaurantsResponseItem,
);

/**
 * @summary Get restaurant by ID
 */
export const GetRestaurantParams = zod.object({
  id: zod.coerce.string(),
});

export const GetRestaurantResponse = zod
  .object({
    id: zod.string(),
    name: zod.string(),
    area: zod.string(),
    address: zod.string(),
    description: zod.string(),
    image_url: zod.string().nullish(),
    avg_score: zod.number().nullish(),
    total_noms: zod.number(),
    price_range: zod.string().nullish(),
    tags: zod.array(zod.string()),
    latitude: zod.number().nullish(),
    longitude: zod.number().nullish(),
    created_at: zod.string(),
  })
  .and(
    zod.object({
      noms: zod
        .array(
          zod.object({
            id: zod.string(),
            restaurant_id: zod.string(),
            user_id: zod.string(),
            user_name: zod.string(),
            user_avatar: zod.string().nullish(),
            score: zod.number(),
            comment: zod.string().nullish(),
            created_at: zod.string(),
          }),
        )
        .optional(),
    }),
  );

/**
 * @summary Get all noms for a restaurant
 */
export const GetRestaurantNomsParams = zod.object({
  id: zod.coerce.string(),
});

export const GetRestaurantNomsResponseItem = zod.object({
  id: zod.string(),
  restaurant_id: zod.string(),
  user_id: zod.string(),
  user_name: zod.string(),
  user_avatar: zod.string().nullish(),
  score: zod.number(),
  comment: zod.string().nullish(),
  created_at: zod.string(),
});
export const GetRestaurantNomsResponse = zod.array(
  GetRestaurantNomsResponseItem,
);

/**
 * @summary Submit a nom score
 */
export const CreateNomBody = zod.object({
  restaurant_id: zod.string(),
  user_id: zod.string(),
  user_name: zod.string(),
  user_avatar: zod.string().nullish(),
  score: zod.number(),
  comment: zod.string().nullish(),
});

/**
 * @summary Get user profile
 */
export const GetUserParams = zod.object({
  id: zod.coerce.string(),
});

export const GetUserResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  email: zod.string(),
  avatar_url: zod.string().nullish(),
  total_noms: zod.number(),
  joined_at: zod.string(),
});

/**
 * @summary Get all noms by a user
 */
export const GetUserNomsParams = zod.object({
  id: zod.coerce.string(),
});

export const GetUserNomsResponseItem = zod.object({
  id: zod.string(),
  restaurant_id: zod.string(),
  user_id: zod.string(),
  user_name: zod.string(),
  user_avatar: zod.string().nullish(),
  score: zod.number(),
  comment: zod.string().nullish(),
  created_at: zod.string(),
});
export const GetUserNomsResponse = zod.array(GetUserNomsResponseItem);

/**
 * @summary Get platform summary stats
 */
export const GetSummaryStatsResponse = zod.object({
  total_restaurants: zod.number(),
  total_noms: zod.number(),
  top_area: zod.string().nullish(),
  avg_score_platform: zod.number().nullish(),
});