# lib/api-zod/src/generated/types/createRestaurantBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Burger Nomr - Halaal burger nom platform for Cape Town
 * OpenAPI spec version: 0.2.0
 */

export interface CreateRestaurantBody {
  name: string;
  area: string;
  address: string;
  description: string;
  image_url?: string | null;
  price_range?: string | null;
  tags: string[];
  latitude?: number | null;
  longitude?: number | null;
}