# lib/api-zod/src/generated/types/nom.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Burger Nomr - Halaal burger nom platform for Cape Town
 * OpenAPI spec version: 0.2.0
 */

export interface Nom {
  id: string;
  restaurant_id: string;
  user_id: string;
  user_name: string;
  user_avatar?: string | null;
  score: number;
  comment?: string | null;
  created_at: string;
}