# lib/api-zod/src/generated/types/restaurantDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Burger Nomr - Halaal burger nom platform for Cape Town
 * OpenAPI spec version: 0.2.0
 */
import type { Nom } from "./nom";
import type { Restaurant } from "./restaurant";

export type RestaurantDetail = Restaurant & {
  noms?: Nom[];
};