# lib/api-zod/src/generated/types/summaryStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Burger Nomr - Halaal burger nom platform for Cape Town
 * OpenAPI spec version: 0.2.0
 */

export interface SummaryStats {
  total_restaurants: number;
  total_noms: number;
  top_area?: string | null;
  avg_score_platform?: number | null;
}