# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Burger Nomr - Halaal burger nom platform for Cape Town
 * OpenAPI spec version: 0.2.0
 */

export interface UserProfile {
  id: string;
  name: string;
  email: string;
  avatar_url?: string | null;
  total_noms: number;
  joined_at: string;
}